dev/prod.rpm: avoid failing anacron jobs during installation
Especially in autobuild (machine recently booted from old template) dev build (takes 55-58 minutes)
anacron job /etc/cron.daily/kataharvester was failing because ckan was not installed completely yet
(%post still running). Not really sure why this happened because according to anacrontab the delay
is 5 minutes and it sounds unlikely that the dev.rpm is already being installed after 5 minutes
(need to install build tools, setup repos and build and install solr first. But anyway it happened
nearly every time recently)

In order to avoid this issue stop crond during installation. (There is still a very small window before
the service stop command is executed, but should it happen the error is not fatal for the
installation)

Interesting enough we had a service crond reload at the end of %post, because of exactly the opposite
problem. Installed cron jobs where not picked up by the daemon. But that was from the time
when our cron jobs where still in crontab. Cron doesn't seem to pick up crontab file changes,
but anacron finds new files in /etc/cron.daily immediately when running.